improved crossing detection
I added criteria for points to get added to the "simplified route", then fixed the end-points for segment intersection to include 0, which otherwise was missing crossing points on the spirograph test route (-fixCrossings).  It's still missing one point so crossing detection suffers from floating point rounding when point falls precisely on the intersecting road.  Fixed error where zig-zag warnings were mm instead of km.  Snapping had a potential infinite loop when snap step was 1 -- I set it to have a minimum of 2, although this should be fixed.  I added a warning when a crossing is found in the simplified route but not the detailed route.  This can happen for really badly formed routes (which happen with my random route generator used for regression testing).  I incremented version number to 0.11.