reduced messages
messages were printed when an action might be taken, not when it was taken.  So if cropping wasn't done, a cropping message was still printed, as a notice to user that the cropping section of the code had been reached.  This was confusing, however.... no cropping was being performed.  So I tried to shift message only to where actions were performed.