updated S/F line placement data
Previously I had been assuming RGT reserved 100 meters at both the start and finish lines for riders to gather.  Experiments on a square course indicate this is actually 60 meters for the start line (including lap courses) and 140 meters at the finish (for point-to-point courses).  The default for-shiftSF was thus changed to -60 (from -100).