added altitude correction on -fixCrossings
When -fixCrossings is specified, code will make sure altitude difference is not between 0 and -crossingHeight spec (default 2 meters), exclusive.  This makes sure level crossings are truly level, and that if there's an overpass, it has a minimum clearance.  Also fixed an issue with 4-point U-turns getting confused by repeated points @ end of loop. (solution is to pop the point, then push it back after U-turn check)