changed -flatten
Previously -flatten took arguments in the sequence: altitude, start position, end position, transition.  The problem with this is it could only flatten to constant altitude.  However, more generally, courses may require "flattening" to a constant slope.  So I changed the arguments.

Now it is start-position, start-altitude, finish-position, finish-altitude, transition.  The default finish altitude is the start altitude.  The default finish point is beyond the end of the course.  The transition STARTS at the boundaries here, which is different than how other ranges work in the code.