changed transition function for -maxRadius
previous maxRadius version used cosine squared transition.  This resulted in the lane shift tending to extend a bit far from the critical point, then angling in.  I changed it to cosine to the 4th power.  This begins the transition back sooner, then the transition is more gradual.  This seems to work better on a test case of tight switchbacks.
Changed version number: 0.30 -> 0.301